[IMP] sign: add top level doc + improve structure
This commit is part of a larger work to improve the documentation's structure. The toctree structure used to add levels to create categories and make the structure visually more understandable.
With the new version of the documentation, plenty of these categories are now redundant.

This commit improves the structure of the "Sign" app:

- it removes a useless "overview" category
- it adds a brief explanation at the "app" level, including links to relevant docs